correctly handle array operations with array values in mongoid
mongo uses the `$each` qualifier to indicate that the value is an array,
which causes mongo to add each array value to the attribute's array. see
http://www.mongodb.org/display/DOCS/Updating#Updating-%24addToSetand%24each
for more info.

since this is a backwards-incompatible change (previous versions added the
entire array to the attribute's array, rather than each of the array's
values), bump the version to 0.2.1.

also replace stickler with gemfury.